requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==5.0.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-silk==5.1.0
djangorestframework==3.15.1
gevent==24.2.1
gprof2dot==2022.7.29
greenlet==3.0.3
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uWSGI==2.0.24
zope.event==5.0
zope.interface==6.1
